Unit test with sinon
