+++ package.json
@@ -8,7 +8,7 @@
     "unit": "mocha",
     "e2e": "python3 -m pytest --rootdir=tests",
     "bidi-server": "npm run build && npm run server-no-build --",
-    "server-no-build": "node ./src/.build/server/index.js",
+    "server-no-build": "node ./src/.build/index.js",
     "build": "npm run prettier && npm run clean && rollup --config src/bidiMapper/rollup.config.js && tsc -b src/tsconfig.json",
     "clean": "rimraf ./src/.build",
     "prettier": "npx prettier --write . "
@@ -24,19 +24,19 @@
     "ws": "^7.4.5"
   },
   "devDependencies": {
-    "@types/rollup": "^0.54.0",
-    "@types/websocket": "^1.0.2",
-    "@types/mocha": "^8.2.2",
-    "@types/ws": "^7.4.4",
-    "@rollup/plugin-typescript": "^8.2.1",
     "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-typescript": "^8.2.1",
+    "@types/mocha": "^8.2.2",
+    "@types/sinon": "^10.0.2",
+    "@types/websocket": "^1.0.2",
+    "@types/ws": "^7.4.4",
     "chai": "^4.3.4",
-    "mocha": "^9.0.1",
     "devtools-protocol": "^0.0.900357",
+    "mocha": "^9.0.1",
     "prettier": "2.3.0",
     "rimraf": "^3.0.2",
     "rollup": "^2.47.0",
-    "typescript": "^4.2.4",
-    "ts-mockito": "^2.6.1"
+    "sinon": "^11.1.1",
+    "typescript": "^4.2.4"
   }
 }
