+++ src/bidiMapper/commandProcessor.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { IServer } from './utils/iServer';
+import { IServer } from './utils/Server/iServer';
 import { BrowsingContextProcessor } from './domains/context/browsingContextProcessor';
 import { Context } from './domains/context/context';
