+++ src/bidiMapper/utils/Server/ServerBinding.ts
@@ -13,15 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/**
- * High-level abstraction of APIs like BiDi or CDP.
- */
-export interface IServer {
-  setOnMessage: (handler: (messageObj: any) => Promise<void>) => void;
-  sendMessage: (messageObj: any) => Promise<any>;
-}
-
 /**
  * Low-level abstraction of API like BiDi or CDP used by IServer implementation.
  */
@@ -30,7 +21,6 @@ export class ServerBinding {
     messageHandler: (messageStr: string) => void
   ) => void;
   private _sendMessage: (messageStr: string) => void;
-
   constructor(
     sendMessage: (messageStr: string) => void,
     messageHandlerSetter: (messageHandler: (messageStr: string) => void) => void
@@ -38,38 +28,10 @@ export class ServerBinding {
     this._messageHandlerSetter = messageHandlerSetter;
     this._sendMessage = sendMessage;
   }
-
-  public set onmessage(messageHandler: (string) => void) {
+  public setOnMessage(messageHandler: (string) => void) {
     this._messageHandlerSetter(messageHandler);
   }
-
   public sendMessage(message: string) {
     this._sendMessage(message);
-  }
-}
-
-export abstract class AbstractServer implements IServer {
-  protected _binding: ServerBinding;
-  private _handlers: ((messageObj: any) => void)[] = new Array();
-
-  constructor(binding: ServerBinding) {
-    this._binding = binding;
-  }
-
-  abstract sendMessage(messageObj: any): Promise<any>;
-
-  /**
-   * Sets handler, which will be called for each CDP message, except commands.
-   * Commands result will be returned by the command promise.
-   * @param handler
-   */
-  setOnMessage(handler: (messageObj: any) => Promise<void>): void {
-    this._handlers.push(handler);
-  }
-
-  protected notifySubscribersOnMessage(messageObj: any): void {
-    for (let handler of this._handlers) {
-      handler(messageObj);
-    }
   }
 }
