+++ src/bidiMapper/utils/Tests/stub.spec.ts
@@ -1,0 +1,24 @@
+/**
+ * Copyright 2021 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { assert, createStubInstance, SinonStubbedInstance } from 'sinon';
+
+export class Stub {
+  static stub<T>(constructor: new (...args) => T): StubbedClass<T> {
+    return createStubInstance(constructor) as any as StubbedClass<T>;
+  }
+  static assert = assert;
+}
+type StubbedClass<T> = SinonStubbedInstance<T> & T;
